Rough prototype of a mesh simplification using accumulated 'memory' quadrics and a collapse priority queue.
Started with memory quadrics as they are much simpler than memoryless, because we don't need to recompute quadrics in a large neighborhood. Edge and wedge quadrics are simply accumulated on vertices and attributes. Even with memory quadric scheme edge errors have to be recomputed in a large radius following an edge collapse. When decimating dragon 3D scan we have to recompute errors for around 50 edges on average and over 200 edges in the worst case.